Revert "Experiment Job refactoring (#237)"
This reverts commit cc383c5c75738655ff6f9625a7438400593dae2b.
