+++ services/python-images/requirements.txt
@@ -15,7 +15,7 @@ Flask==1.0.2
 Flask-Migrate==2.3.1
 Flask-RESTful==0.3.6
 flask-restful-swagger-2==0.35
-Flask-SQLAlchemy==2.4.4
+Flask-SQLAlchemy==2.4.1
 Flask-SocketIO==4.2.1
 idna==2.7
 ijson==2.3
@@ -30,14 +30,14 @@ mccabe==0.6.1
 more-itertools==4.3.0
 networkx==2.2
 numpy==1.16.0
-pandas==1.1.4
+pandas==0.23.4
 pluggy==0.8.0
 psutil==5.4.8
-psycopg2==2.8.6
-psycopg2-binary==2.8.6
+psycopg2==2.7.6.1
+psycopg2-binary==2.7.6.1
 py==1.7.0
 pycodestyle==2.4.0
-pyflakes==2.2.0
+pyflakes==2.0.0
 pyhdb==0.3.4
 pytest==3.10.1
 pytest-ordering==0.6
@@ -45,9 +45,9 @@ python-dateutil==2.7.5
 python-editor==1.0.3
 pytz==2018.7
 requests==2.20.1
-scipy==1.5.4
+scipy==1.2.1
 six==1.11.0
-SQLAlchemy==1.3.20
+SQLAlchemy==1.3.10
 sqlalchemy-hana==0.3.0
 text-unidecode==1.2
 urllib3==1.24.2
