+++ services/python-images/src/master/resources/executor.py
@@ -7,7 +7,7 @@ from flask_restful_swagger_2 import swagger
 from src.db import db
 from src.master.helpers.database import check_dataset_hash
 from src.master.helpers.io import marshal
-from src.models import JobSchema, JobStatus, Experiment, ExperimentJob
+from src.models import Job, JobSchema, JobStatus, Experiment
 
 
 class ExecutorResource(Resource):
@@ -57,11 +57,10 @@ class ExecutorResource(Resource):
             gpus = body.get('gpus')
             enforce_cpus = body.get('enforce_cpus')
         for i in range(runs):
-            new_experiment_job = ExperimentJob(start_time=datetime.now(), status=JobStatus.waiting,
-                                               node_hostname=node_hostname, experiment=experiment,
-                                               parallel=parallel, gpus=gpus, enforce_cpus=enforce_cpus)
-
-            db.session.add(new_experiment_job)
+            new_job = Job(experiment=experiment, start_time=datetime.now(),
+                          status=JobStatus.waiting, node_hostname=node_hostname,
+                          parallel=parallel, gpus=gpus, enforce_cpus=enforce_cpus)
+            db.session.add(new_job)
         db.session.commit()
 
-        return marshal(JobSchema, new_experiment_job)
+        return marshal(JobSchema, new_job)
