+++ services/python-images/src/models/job.py
@@ -23,25 +23,27 @@ class JobErrorCode(int, enum.Enum):
 
 
 class Job(BaseModel):
+    experiment_id = db.Column(
+        db.Integer,
+        db.ForeignKey('experiment.id'),
+        nullable=False
+    )
+    experiment = db.relationship('Experiment', backref=db.backref('jobs', cascade="all, delete-orphan"))
     start_time = db.Column(db.DateTime, nullable=False)
     container_id = db.Column(db.String)
     node_hostname = db.Column(db.String)
     status = db.Column(db.Enum(JobStatus), nullable=False)
     log = db.Column(db.String)
+    parallel = db.Column(db.Boolean)
+    enforce_cpus = db.Column(db.Boolean, default=True)
+    gpus = db.Column(db.Integer)
     error_code = db.Column(db.Enum(JobErrorCode))
-
-    type = db.Column(db.String)
 
     @property
     def result(self):
         if len(self.results) == 0:
             return None
         return sorted(self.results, key=lambda x: x.start_time)[-1]
-
-    __mapper_args__ = {
-        'polymorphic_identity': 'job',
-        'polymorphic_on': type
-    }
 
 
 class JobSchema(BaseSchema):
